+++ turbinia/client_gcp_test.py
@@ -21,13 +21,13 @@ from datetime import timedelta
 import json
 import unittest
 import textwrap
-
+import importlib
 import mock
 
 from turbinia import config
 from turbinia import client as TurbiniaClientProvider
+from turbinia import state_manager
 from turbinia import TurbiniaException
-from turbinia.client import TurbiniaCeleryClient, TurbiniaStats
 
 DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%S.%fZ'
 
@@ -217,8 +217,22 @@ STATISTICS_REPORT_CSV = textwrap.dedent(
 class TestTurbiniaClient(unittest.TestCase):
   """Test Turbinia client class."""
 
-  def setUp(self):
+  def _get_state_manager(self):
+    """Gets a Datastore State Manager object for test."""
+    config.STATE_MANAGER = 'Datastore'
+    config.TASK_MANAGER = 'PSQ'
+    config.TURBINIA_PROJECT = 'test'
+    config.TURBINIA_ZONE = 'test'
+    config.TURBINIA_REGION = 'test'
+    # Reload module using the config settings above.
+    # This is needed due to the conditional imports in client.py
+    importlib.reload(TurbiniaClientProvider)
     TurbiniaClientProvider.RETRY_SLEEP = 0.001
+    return state_manager.get_state_manager()
+
+  def setUp(self):
+    self.state_manager = self._get_state_manager()
+
     last_update = datetime.strptime(
         '2020-08-04T16:32:38.390390Z', DATETIME_FORMAT)
     self.task_data = [
@@ -272,82 +286,96 @@ class TestTurbiniaClient(unittest.TestCase):
 
 
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testTurbiniaClientInit(self, _, __):
     """Basic test for client."""
     config.LoadConfig()
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     self.assertTrue(hasattr(client, 'task_manager'))
 
-  @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testTurbiniaClientGetTaskData(self, _, __, mock_cloud_function):
+  @mock.patch('turbinia.state_manager.datastore.Client')
+  @mock.patch(
+      'libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction'
+  )
+  def testTurbiniaClientGetTaskData(self, mock_cloud_function, _, __):
     """Basic test for client.get_task_data"""
-    if config.STATE_MANAGER.lower() == 'datastore':
-      # ExecuteFunction returns a dict with a 'result' key that has a json-encoded
-      # list.  This contains our task data, which is a list of dicts.
-      run_time = timedelta(seconds=3)
-      test_task_data = [{'bar': 'bar2', 'run_time': run_time.total_seconds()}]
-      gcf_result = [test_task_data, 'Unused GCF data']
-      gcf_result = json.dumps(gcf_result)
-      function_return = {'result': gcf_result}
-      mock_cloud_function.return_value = function_return
-      client = TurbiniaClientProvider.get_turbinia_client()
-      task_data = client.get_task_data('inst', 'proj', 'reg')
-      # get_task_data() converts this back into a timedelta(). We returned it
-      # seconds from the GCF function call because that is what it is stored in
-      # Datastore as.
-      test_task_data[0]['run_time'] = run_time
-      self.assertEqual(task_data, test_task_data)
+    # ExecuteFunction returns a dict with a 'result' key that has a json-encoded
+    # list.  This contains our task data, which is a list of dicts.
+    run_time = timedelta(seconds=3)
+    test_task_data = [{'bar': 'bar2', 'run_time': run_time.total_seconds()}]
+    gcf_result = [test_task_data, 'Unused GCF data']
+    gcf_result = json.dumps(gcf_result)
+    function_return = {'result': gcf_result}
+    mock_cloud_function.return_value = function_return
+    client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
+    task_data = client.get_task_data('inst', 'proj', 'reg')
+    # get_task_data() converts this back into a timedelta(). We returned it
+    # seconds from the GCF function call because that is what it is stored in
+    # Datastore as.
+    test_task_data[0]['run_time'] = run_time
+    self.assertEqual(task_data, test_task_data)
 
-      # Also test that JSON output works
-      task_data = client.get_task_data('inst', 'proj', 'reg', output_json=True)
-      self.assertEqual(task_data, '[{"bar": "bar2", "run_time": 3.0}]')
+    # Also test that JSON output works
+    task_data = client.get_task_data('inst', 'proj', 'reg', output_json=True)
+    self.assertEqual(task_data, '[{"bar": "bar2", "run_time": 3.0}]')
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
+  @mock.patch(
+      'libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction')
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testTurbiniaClientGetTaskDataNoResults(self, _, __, mock_cloud_function):
+  @mock.patch('turbinia.state_manager.datastore.Client')
+  def testTurbiniaClientGetTaskDataNoResults(
+      self, _, __, ___, mock_cloud_function):
     """Test for exception after empty results from cloud functions."""
-    if config.STATE_MANAGER.lower() == 'datastore':
-      mock_cloud_function.return_value = {}
-      client = TurbiniaClientProvider.get_turbinia_client()
-      self.assertRaises(
-          TurbiniaException, client.get_task_data, "inst", "proj", "reg")
+    mock_cloud_function.return_value = {}
+    client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
+    self.assertRaises(
+        TurbiniaException, client.get_task_data, "inst", "proj", "reg")
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testTurbiniaClientGetTaskDataRetriableErrors(
       self, _, __, mock_cloud_function):
     """Test for retries after retriable errors returned from cloud functions."""
-    if config.STATE_MANAGER.lower() == 'datastore':
-      mock_cloud_function.return_value = {'error': {'code': 503}}
-      client = TurbiniaClientProvider.get_turbinia_client()
-      self.assertRaises(
-          TurbiniaException, client.get_task_data, "inst", "proj", "reg")
-      self.assertEqual(
-          mock_cloud_function.call_count, TurbiniaClientProvider.MAX_RETRIES)
+    mock_cloud_function.return_value = {'error': {'code': 503}}
+    client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
+    self.assertRaises(
+        TurbiniaException, client.get_task_data, "inst", "proj", "reg")
+    self.assertEqual(
+        mock_cloud_function.call_count, TurbiniaClientProvider.MAX_RETRIES)
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
+  @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testTurbiniaClientGetTaskDataInvalidJson(
-      self, _, __, mock_cloud_function):
+      self, _, __, ___, mock_cloud_function):
     """Test for exception after bad json results from cloud functions."""
-    if config.STATE_MANAGER.lower() == 'datastore':
-      mock_cloud_function.return_value = {'result': None}
-      client = TurbiniaClientProvider.get_turbinia_client()
-      self.assertRaises(
-          TurbiniaException, client.get_task_data, "inst", "proj", "reg")
+    mock_cloud_function.return_value = {'result': None}
+    client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
+    self.assertRaises(
+        TurbiniaException, client.get_task_data, "inst", "proj", "reg")
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatTaskStatistics(self, _, __, ___):
     """Tests format_task_statistics() report output."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     stats_report = client.format_task_statistics('inst', 'proj', 'reg')
@@ -356,10 +384,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatTaskStatisticsCsv(self, _, __, ___):
     """Tests format_task_statistics() CSV report output."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     stats_report = client.format_task_statistics(
@@ -369,10 +399,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientGetTaskStatistics(self, _, __, ___):
     """Tests get_task_statistics() basic functionality."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     task_stats = client.get_task_statistics('inst', 'proj', 'reg')
@@ -402,10 +434,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testClientFormatTaskStatus(self, _, __, ___):
+  @mock.patch('turbinia.state_manager.datastore.Client')
+  def testClientFormatTaskStatus(self, _, __, ____):
     """Tests format_task_status() with empty report_priority."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     self.task_data[0]['report_priority'] = None
     self.task_data[1]['report_priority'] = ''
@@ -416,10 +450,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatTaskStatusShortReport(self, _, __, ___):
     """Tests format_task_status() has valid output with short report."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_task_status('inst', 'proj', 'reg')
@@ -427,10 +463,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testClientFormatTaskStatusFullReport(self, _, __, ___):
+  @mock.patch('turbinia.state_manager.datastore.Client')
+  def testClientFormatTaskStatusFullReport(self, _, __, mock_cloud_function):
     """Tests format_task_status() has valid output with full report."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_task_status('inst', 'proj', 'reg', full_report=True)
@@ -438,10 +476,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
-  def testClientFormatTaskStatusFiles(self, _, __, ___):
+  @mock.patch('turbinia.state_manager.datastore.Client')
+  def testClientFormatTaskStatusFiles(self, _, __, mock_cloud_function):
     """Tests format_task_status() has valid output with report and files."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_task_status(
@@ -450,10 +490,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatRequestStatus(self, _, __, ___):
     """Tests format_request_status() with default days."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_request_status('inst', 'proj', 'reg')
@@ -461,10 +503,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatRequestStatusDays(self, _, __, ___):
     """Tests format_request_status() with custom days."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_request_status('inst', 'proj', 'reg', days=4)
@@ -472,10 +516,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatRequestStatusNoResults(self, _, __, ___):
     """Tests format_request_status() with no Task results."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = ''
     result = client.format_request_status('inst', 'proj', 'reg', days=4)
@@ -483,10 +529,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatRequestStatusFullReport(self, _, __, ___):
     """Tests format_request_status() has valid output with full report."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_request_status(
@@ -495,10 +543,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatWorkerStatus(self, _, __, ___):
     """Tests format_worker_status() with default days."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_worker_status('inst', 'proj', 'reg')
@@ -507,10 +557,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatWorkerStatusDays(self, _, __, ___):
     """Tests format_worker_status() with custom days."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_worker_status('inst', 'proj', 'reg', days=4)
@@ -519,10 +571,12 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatWorkerStatusNoResults(self, _, __, ___):
     """Tests format_worker_status() with no Task results."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = ''
     result = client.format_worker_status('inst', 'proj', 'reg', days=4)
@@ -530,76 +584,13 @@ class TestTurbiniaClient(unittest.TestCase):
 
   @mock.patch('libcloudforensics.providers.gcp.internal.function.GoogleCloudFunction.ExecuteFunction')  # yapf: disable
   @mock.patch('turbinia.client.task_manager.PSQTaskManager._backend_setup')
-  @mock.patch('turbinia.state_manager.get_state_manager')
+  @mock.patch('turbinia.state_manager.datastore.Client')
   def testClientFormatWorkStatusFullReport(self, _, __, ___):
     """Tests format_worker_status() has valid output with full report."""
     client = TurbiniaClientProvider.get_turbinia_client()
+    client.task_amanger = self.state_manager
+    #client.task_manager = self._get_state_manager()
     client.get_task_data = mock.MagicMock()
     client.get_task_data.return_value = self.task_data
     result = client.format_worker_status('inst', 'proj', 'reg', all_fields=True)
     self.assertEqual(result.strip(), LONG_REPORT_WORKERS.strip())
-
-
-class TestTurbiniaStats(unittest.TestCase):
-  """Test TurbiniaStats class."""
-
-  def testTurbiniaStatsAddTask(self):
-    """Tests TurbiniaStats.add_task() method."""
-    test_task = {'run_time': None, 'last_update': None}
-    stats = TurbiniaStats()
-    stats.add_task(test_task)
-    self.assertIn(test_task, stats.tasks)
-    self.assertEqual(stats.count, 1)
-
-  def testTurbiniaStatsCalculateStats(self):
-    """Tests TurbiniaStats.calculateStats() method."""
-    last_update = datetime.now()
-    test_task1 = {'run_time': timedelta(minutes=3), 'last_update': last_update}
-    test_task2 = {'run_time': timedelta(minutes=5), 'last_update': last_update}
-    test_task3 = {'run_time': timedelta(minutes=1), 'last_update': last_update}
-
-    stats = TurbiniaStats()
-    stats.add_task(test_task1)
-    stats.add_task(test_task2)
-    stats.add_task(test_task3)
-    stats.calculate_stats()
-
-    self.assertEqual(stats.min, timedelta(minutes=1))
-    self.assertEqual(stats.mean, timedelta(minutes=3))
-    self.assertEqual(stats.max, timedelta(minutes=5))
-    self.assertEqual(stats.count, 3)
-
-  def testTurbiniaStatsCalculateStatsEmpty(self):
-    """Tests that calculate_stats() works when no tasks are added."""
-    stats = TurbiniaStats()
-    stats.calculate_stats()
-    self.assertEqual(stats.count, 0)
-    self.assertEqual(stats.min, None)
-
-  def testTurbiniaStatsFormatStats(self):
-    """Tests TurbiniaStats.format_stats() returns valid output."""
-    test_output = (
-        'Test Task Results: Count: 1, Min: 0:03:00, Mean: 0:03:00, '
-        'Max: 0:03:00')
-    test_task1 = {
-        'run_time': timedelta(minutes=3),
-        'last_update': datetime.now()
-    }
-    stats = TurbiniaStats('Test Task Results')
-    stats.add_task(test_task1)
-    stats.calculate_stats()
-    report = stats.format_stats()
-    self.assertEqual(report, test_output)
-
-  def testTurbiniaStatsFormatStatsCsv(self):
-    """Tests TurbiniaStats.format_stats() returns valid CSV output."""
-    test_output = ('Test Task Results, 1, 0:03:00, 0:03:00, 0:03:00')
-    test_task1 = {
-        'run_time': timedelta(minutes=3),
-        'last_update': datetime.now()
-    }
-    stats = TurbiniaStats('Test Task Results')
-    stats.add_task(test_task1)
-    stats.calculate_stats()
-    report = stats.format_stats_csv()
-    self.assertEqual(report, test_output)
